refactor: move extension handling to custom hook
